# openapi.yaml
openapi: 3.0.0
info:
  title: GPT Custom Server API
  version: 1.0.0
  description: This API allows a custom GPT instance to interact with files and run Jest tests programmatically.
servers:
  - url: https://everydaywetunnelin.ngrok.io
    description: ngrok public URL
    variables:
      subdomain:
        default: your_subdomain
paths:
  /file:
    get:
      summary: Fetch the contents of the file
      operationId: getFileContents
      parameters:
        - in: query
          name: filePath
          required: true
          schema:
            type: string
          description: The path of the file to fetch
      responses:
        "200":
          description: File contents retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                    example: "test('simple test', () => { expect(true).toBe(false); });"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "File read error message"
    post:
      summary: Send the file with changes, apply changes locally, run Jest tests, and return the test results
      operationId: postFileChanges
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                  example: "example.spec.js"
                content:
                  type: string
                  example: "test('simple test', () => { expect(true).toBe(true); });"
      responses:
        "200":
          description: Test results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: string
                    example: "PASS  ./example.spec.ts\n  simple test\n    ✓ simple test (2 ms)"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Test run error message"
  /all-files:
    get:
      summary: Run the `npm run all` command and return the output
      operationId: getAllFiles
      parameters:
        - in: query
          name: directory
          schema:
            type: string
          description: The directory to list the files from
      responses:
        "200":
          description: Command output retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: object
                    properties:
                      allFileContents:
                        type: string
                        example: "Directory structure:\n\n├── Dockerfile\n├── docker-compose.yaml\n..."
                      fileTree:
                        type: string
                        example: "├── Dockerfile\n├── docker-compose.yaml\n..."
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Command execution error message"
  /all-tests:
    get:
      summary: Run all Jest tests and return the output
      operationId: getAllTests
      responses:
        "200":
          description: Test output retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: string
                    example: "PASS  ./example.spec.js\n  simple test\n    ✓ simple test (2 ms)"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Test run error message"
  /commit-files:
    post:
      summary: Commit file changes with a commit message, run tests, and return the results
      operationId: commitFiles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                commitMessage:
                  type: string
                  example: "Updated test case to expect true"
                files:
                  type: array
                  items:
                    type: object
                    properties:
                      filePath:
                        type: string
                        example: "example.spec.js"
                      content:
                        type: string
                        example: "test('simple test', () => { expect(true).toBe(true); });"
      responses:
        "200":
          description: Test results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: string
                    example: "PASS  ./example.spec.js\n  simple test\n    ✓ simple test (2 ms)"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Commit and test run error message"
  /select-files:
    post:
      summary: Select a set of files to retrieve later
      operationId: selectFiles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                  example: ["Dockerfile", "server.js", "scripts/outputRepoContents.js"]
      responses:
        "200":
          description: Files selected successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Files selected successfully"
                  selectedFiles:
                    type: array
                    items:
                      type: string
                    example: ["Dockerfile", "server.js", "scripts/outputRepoContents.js"]
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Files should be an array"
  /selected-files:
    get:
      summary: Retrieve the contents of selected files
      operationId: getSelectedFiles
      responses:
        "200":
          description: Selected files contents retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  allFileContents:
                    type: string
                    example: "Contents of selected files:\n\n### Dockerfile ###\n\n...file content...\n\n### server.js ###\n\n...file content...\n\n### scripts/outputRepoContents.js ###\n\n...file content...\n\n"
        "400":
          description: No files selected
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No files selected"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "File read error message"
components:
  schemas: {}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []